changed some ints to floats for damage

+++ Assets/Scripts/EnemyAI.cs
@@ -13,7 +13,7 @@ public class EnemyAI : MonoBehaviour, IDamage
     [SerializeField] Animator animator;
 
     [Header("---Enemy Stats---")]
-    [SerializeField] int HP;
+    [SerializeField] float HP;
     [SerializeField] int playerFaceSpeed;
     [SerializeField] int sightAngle;
     [SerializeField] Transform headPos;
@@ -27,7 +27,7 @@ public class EnemyAI : MonoBehaviour, IDamage
     [SerializeField] Image enemyHPBar;
     [SerializeField] GameObject enemyUI;
 
-    int HPorg;
+    float HPorg;
     bool isShooting;
     bool playerInRange;
     Vector3 playerDirection;
@@ -125,7 +125,7 @@ public class EnemyAI : MonoBehaviour, IDamage
         enemyHPBar.fillAmount = (float)HP / (float)HPorg;
 
     }
-    public void takeDamage(int dmg)
+    public void takeDamage(float dmg)
     {
         HP -= dmg;
         UpdateEnemyHPBar();
@@ -140,7 +140,7 @@ public class EnemyAI : MonoBehaviour, IDamage
                 Instantiate(enemyDrop, shootPos.position, transform.rotation);
             }
             StartCoroutine(Death());
-            GameManager.instance.playerScript.AddCoins(HPorg);
+            GameManager.instance.playerScript.AddCoins((int)HPorg);
             GameManager.instance.UpdateEnemyCount(-1);
             
         }
